feat: dto 수정 및 detail 네비게이션 연결

+++ app/src/main/java/com/android/quizcafe/core/data/model/quizbook/request/QuizBookDetailRequestDto.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class QuizBookDetailRequestDto(
-    val quizBookId: Int
+    val quizBookId: Long
 )
 
 fun QuizBookDetailRequest.toDto() =
