+++ app/src/main/java/com/android/quizcafe/core/data/remote/datasource/QuizBookRemoteDataSource.kt
@@ -24,9 +24,9 @@ class QuizBookRemoteDataSource @Inject constructor(
     suspend fun getQuizBooksByCategory(request: QuizBookRequestDto): NetworkResult<ApiResponse<List<QuizBookResponseDto>>> =
         quizBookService.getQuizBooks(request.category)
 
-    suspend fun saveQuizBook(quizBookId: Int): NetworkResult<ApiResponse<Unit>> =
+    suspend fun saveQuizBook(quizBookId: Long): NetworkResult<ApiResponse<Unit>> =
         quizBookService.saveQuizBook(quizBookId)
 
-    suspend fun unsaveQuizBook(quizBookId: Int): NetworkResult<ApiResponse<Unit>> =
+    suspend fun unsaveQuizBook(quizBookId: Long): NetworkResult<ApiResponse<Unit>> =
         quizBookService.unsaveQuizBook(quizBookId)
 }
