+++ app/src/main/java/com/android/quizcafe/core/domain/model/quizbook/request/QuizBookDetailRequest.kt
@@ -2,5 +2,5 @@ package com.android.quizcafe.core.domain.model.quizbook.request
 
 // TODO: API Response에 맞게 수정
 data class QuizBookDetailRequest(
-    val quizBookId: Int
+    val quizBookId: Long
 )
